Assignment-1/requirements.txt
snap-stanford==5.0.0
